Refactor to consolidate dashboard tests
